backend/requirements.txt
Django==5.0.2
channels==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
# For API testing
requests==2.31.0
websocket-client==1.7.0 